Warning about future changes
It is worth noting that the utility of this script depends on the accuracy of it's template string. If files are re-QC'd, or something else happens in future which changes the dates of these directories, this will effectively create broken output. Be wary of this (Code commented with similar message)