Added example input file from IMPC
The column name and ordering for this file is: ['mp_id',	'id_link',	'mp_term',	'mp_definition',	'mgi_accession_id',	'marker_symbol',	'human_gene_symbol']
Of these, mp_id, human_gene_symbol are used to obtain the lists. The genes in these columns are concatenated with '|', so this is how they are separated into a list within the batch_process script. The mp_ids could be used to separate these into different files (like with the park genes)